Update example to use defn instead of def
Using def makes the data structure be compiled only once, which is fine if the vals aren't changing within the structure, but if they are, calling a `reframe/subscribe` on a value to be inserted into the table-structure causes it to fail, as the value is updated more than once. See https://stackoverflow.com/questions/4461320/clojure-def-vs-defn-for-a-function-with-no-arguments. The following is an example that uses defn correctly for Reformation. 
```(defn test-form []
  (let [default-vals (rfc/subscribe [:get-edit-form :devices])]
    [:devices [:edit
               [:myhidden-text {:type :hidden
                                :default-value "whisper"}
                :model {:type :text
                        :default-value (:model @default-vals)
                        :required? true
                        :style-classes "has-text-centered"
                        :label "Model"}]]]))